Refactor item adding dialog
- Remove menu button from dialog start
- Move dialog routes to separate modules in add_item package
